# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * RoVerse Dashboard API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  admin: 'admin',
  collaborator: 'collaborator',
} as const;

export interface User {
  id: number;
  robloxId: string;
  robloxUsername: string;
  /** @nullable */
  robloxDisplayName?: string | null;
  /** @nullable */
  robloxAvatarUrl?: string | null;
  role: UserRole;
  subroles?: string[];
  groups?: string[];
  createdAt: string;
}

export type UserInputRole = typeof UserInputRole[keyof typeof UserInputRole];


export const UserInputRole = {
  admin: 'admin',
  collaborator: 'collaborator',
} as const;

export interface UserInput {
  robloxId: string;
  role: UserInputRole;
  subroles?: string[];
  groups?: string[];
}

export type UserUpdateRole = typeof UserUpdateRole[keyof typeof UserUpdateRole];


export const UserUpdateRole = {
  admin: 'admin',
  collaborator: 'collaborator',
} as const;

export interface UserUpdate {
  role?: UserUpdateRole;
  subroles?: string[];
  groups?: string[];
}

export interface Game {
  id: number;
  robloxGameId: string;
  name: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  gameLink?: string | null;
  /** @nullable */
  groupLink?: string | null;
  /** @nullable */
  thumbnailUrl?: string | null;
  isActive?: boolean;
  createdAt: string;
}

export interface GameInput {
  robloxGameId: string;
  name: string;
  description?: string;
  gameLink?: string;
  groupLink?: string;
}

export interface GameUpdate {
  name?: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  gameLink?: string | null;
  /** @nullable */
  groupLink?: string | null;
  isActive?: boolean;
}

export interface TimeseriesPoint {
  timestamp: string;
  value: number;
}

export interface GameStats {
  gameId: number;
  ccu: number;
  peakCcu?: number;
  /** In Robux */
  avgDailyRevenue: number;
  /** In Euros */
  avgDailyRevenueEur: number;
  monthlyEstimateRobux?: number;
  monthlyEstimateEur?: number;
  revenuePercentChange?: number;
  ccuHistory?: TimeseriesPoint[];
}

export interface DashboardSummary {
  totalGames: number;
  totalCcu: number;
  totalDailyRevenue: number;
  totalDailyRevenueEur: number;
  monthlyEstimateRobux: number;
  monthlyEstimateEur: number;
  topGame?: Game;
}

export interface RevenueTrackerEntry {
  gameId: number;
  gameName: string;
  /** Percentage of total revenue */
  revenuePercentage: number;
  monthlyEstimateRobux: number;
  monthlyEstimateEur: number;
  /** Percentage change vs last month */
  trend: number;
}

export type TeamMemberRole = typeof TeamMemberRole[keyof typeof TeamMemberRole];


export const TeamMemberRole = {
  admin: 'admin',
  collaborator: 'collaborator',
} as const;

export interface TeamMember {
  userId: number;
  username: string;
  /** @nullable */
  displayName?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  role: TeamMemberRole;
}

export interface Team {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  color: string;
  createdAt: string;
  members: TeamMember[];
}

export interface TeamInput {
  name: string;
  description?: string;
  color?: string;
}

export interface TeamUpdate {
  name?: string;
  /** @nullable */
  description?: string | null;
  color?: string;
}

export interface AddTeamMemberInput {
  userId: number;
}

export interface Board {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  color?: string | null;
  /** @nullable */
  teamId?: number | null;
  /** @nullable */
  teamName?: string | null;
  createdAt: string;
}

export interface BoardInput {
  name: string;
  description?: string;
  color?: string;
  teamId?: number;
}

export interface BoardUpdate {
  name?: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  color?: string | null;
}

export interface BoardColumn {
  id: number;
  boardId: number;
  name: string;
  position: number;
}

/**
 * @nullable
 */
export type TaskPriority = typeof TaskPriority[keyof typeof TaskPriority] | null;


export const TaskPriority = {
  low: 'low',
  medium: 'medium',
  high: 'high',
} as const;

export interface Task {
  id: number;
  boardId: number;
  columnId: number;
  title: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  assigneeId?: number | null;
  /** @nullable */
  assigneeUsername?: string | null;
  /** @nullable */
  priority?: TaskPriority;
  /** @nullable */
  dueDate?: string | null;
  position: number;
  tags?: string[];
  createdAt: string;
}

export interface ColumnWithTasks {
  id: number;
  boardId: number;
  name: string;
  position: number;
  tasks: Task[];
}

export interface BoardDetail {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  color?: string | null;
  columns: ColumnWithTasks[];
  createdAt: string;
}

export type TaskInputPriority = typeof TaskInputPriority[keyof typeof TaskInputPriority];


export const TaskInputPriority = {
  low: 'low',
  medium: 'medium',
  high: 'high',
} as const;

export interface TaskInput {
  boardId: number;
  columnId: number;
  title: string;
  description?: string;
  assigneeId?: number;
  priority?: TaskInputPriority;
  dueDate?: string;
  tags?: string[];
}

/**
 * @nullable
 */
export type TaskUpdatePriority = typeof TaskUpdatePriority[keyof typeof TaskUpdatePriority] | null;


export const TaskUpdatePriority = {
  low: 'low',
  medium: 'medium',
  high: 'high',
} as const;

export interface TaskUpdate {
  columnId?: number;
  title?: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  assigneeId?: number | null;
  /** @nullable */
  priority?: TaskUpdatePriority;
  /** @nullable */
  dueDate?: string | null;
  position?: number;
  tags?: string[];
}

export interface CalendarEvent {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  startDate: string;
  /** @nullable */
  endDate?: string | null;
  /** @nullable */
  color?: string | null;
  /** @nullable */
  assigneeId?: number | null;
  allDay?: boolean;
  createdAt: string;
}

export interface CalendarEventInput {
  title: string;
  description?: string;
  startDate: string;
  endDate?: string;
  color?: string;
  assigneeId?: number;
  allDay?: boolean;
}

export interface CalendarEventUpdate {
  title?: string;
  /** @nullable */
  description?: string | null;
  startDate?: string;
  /** @nullable */
  endDate?: string | null;
  /** @nullable */
  color?: string | null;
  /** @nullable */
  assigneeId?: number | null;
  allDay?: boolean;
}

export interface ColumnInput {
  name: string;
  position?: number;
}

export interface ColumnUpdate {
  name?: string;
  position?: number;
}

export interface RobloxGame {
  universeId: number;
  /** @nullable */
  rootPlaceId?: number | null;
  name: string;
  /** @nullable */
  description?: string | null;
  playing?: number;
  visits?: number;
  favoritedCount?: number;
  isPrivate?: boolean;
  /** @nullable */
  thumbnailUrl?: string | null;
}

export interface GroupInfo {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  memberCount: number;
  /** @nullable */
  thumbnailUrl?: string | null;
  games: RobloxGame[];
}

export type ListTasksParams = {
boardId?: number;
};

export type ListEventsParams = {
from?: string;
to?: string;
};