# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: RoVerse Dashboard API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /games:
    get:
      operationId: listGames
      tags: [games]
      summary: List all linked Roblox games
      responses:
        "200":
          description: List of games
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Game"

    post:
      operationId: createGame
      tags: [games]
      summary: Link a Roblox game
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GameInput"
      responses:
        "201":
          description: Game linked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Game"
  /games/{id}:
    get:
      operationId: getGame
      tags: [games]
      summary: Get a specific game
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Game details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Game"
        "404":
          description: Not found

    patch:
      operationId: updateGame
      tags: [games]
      summary: Update a game
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GameUpdate"
      responses:
        "200":
          description: Updated game
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Game"

    delete:
      operationId: deleteGame
      tags: [games]
      summary: Unlink a game
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /games/{id}/stats:
    get:
      operationId: getGameStats
      tags: [games]
      summary: Get real-time stats for a game (CCU, revenue)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Game stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GameStats"
...
components:
  schemas:
...
    Game:
      type: object
      required: [id, robloxGameId, name, createdAt]
      properties:
        id:
          type: integer
        robloxGameId:
          type: string
        name:
          type: string
        description:
          type: ["string", "null"]
        gameLink:
          type: ["string", "null"]
        groupLink:
          type: ["string", "null"]
        thumbnailUrl:
          type: ["string", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
    GameInput:
      type: object
      required: [robloxGameId, name]
      properties:
        robloxGameId:
          type: string
        name:
          type: string
        description:
          type: string
        gameLink:
          type: string
        groupLink:
          type: string
    GameUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
        gameLink:
          type: ["string", "null"]
        groupLink:
          type: ["string", "null"]
        isActive:
          type: boolean
    GameStats:
      type: object
      required: [gameId, ccu, avgDailyRevenue, avgDailyRevenueEur]
      properties:
        gameId:
          type: integer
        ccu:
          type: integer
        peakCcu:
          type: integer
        avgDailyRevenue:
          type: number
          description: In Robux
        avgDailyRevenueEur:
          type: number
          description: In Euros
        monthlyEstimateRobux:
          type: number
        monthlyEstimateEur:
          type: number
        revenuePercentChange:
          type: number
        ccuHistory:
          type: array
          items:
            $ref: "#/components/schemas/TimeseriesPoint"
    TimeseriesPoint:
      type: object
      required: [timestamp, value]
      properties:
        timestamp:
          type: string
          format: date-time
        value:
          type: number
...